+++ src/main/java/org/sbb/sbb/common/exception/DataNotFoundException.java
@@ -1,0 +1,13 @@
+package org.sbb.sbb.common.config.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "entity not found")
+public class DataNotFoundException extends RuntimeException{
+
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+
+}
