+++ src/main/java/org/sbb/sbb/common/util/CommonUtil.java
