--- src/main/java/org/sbb/sbb/config/auth/LoginService.java
@@ -1,23 +1,0 @@
-package org.sbb.sbb.config.auth;
-
-import lombok.RequiredArgsConstructor;
-import org.sbb.sbb.user.domain.User;
-import org.sbb.sbb.user.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class LoginService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username).orElseThrow(()-> new UsernameNotFoundException("인증실패"));
-
-        return new LoginUser(user);
-    }
-}
