--- src/main/java/org/sbb/sbb/config/auth/LoginUser.java
@@ -1,55 +1,0 @@
-package org.sbb.sbb.config.auth;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.sbb.sbb.user.domain.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-@Getter
-@RequiredArgsConstructor
-public class LoginUser implements UserDetails {
-
-    private final User user;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> authorities = new ArrayList<GrantedAuthority>();
-        authorities.add(() -> user.getRole().getValue());
-        return authorities;
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
