+++ src/main/java/org/sbb/sbb/domain/user/entity/Role.java
@@ -1,0 +1,12 @@
+package org.sbb.sbb.user.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Role {
+    ROLE_USER("유저"), ROLE_ADMIN("관리자");
+
+    private String value;
+}
