+++ src/main/java/org/sbb/sbb/domain/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package org.sbb.sbb.user.repository;
+
+import org.sbb.sbb.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User,Long> {
+
+    Optional<User> findByUsername(String username);
+}
