--- src/main/java/org/sbb/sbb/handler/exception/CustomDataNotFoundException.java
@@ -1,9 +1,0 @@
-package org.sbb.sbb.handler.exception;
-
-public class CustomDataNotFoundException extends RuntimeException{
-
-    public CustomDataNotFoundException(String message) {
-        super(message);
-    }
-
-}
