--- src/main/java/org/sbb/sbb/user/controller/UserController.java
@@ -1,56 +1,0 @@
-package org.sbb.sbb.user.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.sbb.sbb.user.domain.dto.UserReqDto.*;
-import org.sbb.sbb.user.service.UserService;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Controller;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/user")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    @GetMapping("/login")
-    public String login() {
-        return "login_form";
-    }
-
-    @GetMapping("/signup")
-    public String signup(UserJoinReqDto userJoinReqDto) {
-        return "signup_form";
-    }
-
-    @PostMapping("/signup")
-    public String signup(@Valid UserJoinReqDto userJoinReqDto, BindingResult bindingResult) {
-
-        if (bindingResult.hasErrors()) {
-            return "signup_form";
-        }
-
-        if (!userJoinReqDto.getPassword().equals(userJoinReqDto.getPasswordConfirm())) {
-            bindingResult.rejectValue("passwordConfirm", "passwordInCorrect",
-                    "2개의 패스워드가 일치하지 않습니다.");
-            return "signup_form";
-        }
-
-        try {
-            userService.userJoin(userJoinReqDto);
-        }catch(DataIntegrityViolationException e) {
-            bindingResult.reject("signupFailed", "이미 등록된 사용자입니다.");
-            return "signup_form";
-        }catch(Exception e) {
-            bindingResult.reject("signupFailed", e.getMessage());
-            return "signup_form";
-        }
-        return "redirect:/question/list";
-    }
-
-}
