--- src/main/java/org/sbb/sbb/user/domain/User.java
@@ -1,52 +1,0 @@
-package org.sbb.sbb.user.domain;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.sbb.sbb.board.answer.domain.Answer;
-import org.sbb.sbb.board.question.domain.Question;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Table(name = "users")
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String username;
-
-    private String password;
-
-    @Column(unique = true)
-    private String nickname;
-
-    @Column(unique = true)
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-    @CreatedDate
-    private LocalDateTime createDate;
-
-    @LastModifiedDate
-    private LocalDateTime modifyDate;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private List<Question> questionList;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private List<Answer> answerList;
-
-}
