--- src/main/java/org/sbb/sbb/user/domain/dto/UserRespDto.java
@@ -1,4 +1,0 @@
-package org.sbb.sbb.user.domain.dto;
-
-public class UserRespDto {
-}
