--- src/main/java/org/sbb/sbb/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package org.sbb.sbb.user.repository;
-
-import org.sbb.sbb.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-
-    Optional<User> findByUsername(String username);
-}
