+++ src/main/resources/templates/password_update.html
@@ -1,0 +1,16 @@
+<html layout:decorate="~{layout}" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<div layout:fragment="content" class="container my-3">
+    <form th:action="@{/user/update/pw}" th:object="${updatePw}" method="post">
+        <div th:replace="~{form_errors :: formErrorsFragment}"></div>
+        <div class="mb-3">
+            <label th:for="oldPassword" class="form-label">기존 비밀번호</label>
+            <input type="text" th:field="*{oldPassword}" class="form-control">
+        </div>
+        <div class="mb-3">
+            <label th:for="newPassword" class="form-label">새로운 비밀번호</label>
+            <input type="text" th:field="*{newPassword}" class="form-control">
+        </div>
+        <button type="submit" class="btn btn-primary">비밀번호 변경</button>
+    </form>
+</div>
+</html>
